pyproject.toml
[tool.poetry]
name = "donwload sieg hub"
version = "0.1.0"
description = "Download de documentos fiscais do Sieg Hub"
authors = ["frzanini <frzanini@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
nfelib = "^2.1.1"
lxml = "^5.3.0"
xmltodict = "^0.14.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
